+++ src/modules/settings.ts
@@ -33,13 +33,20 @@ const BUTTON_GAP = 20;
 const BUTTON_TEXT_PADDING = 20;
 
 // Player only buttons
-const WIKI_POSITION = [100, 605, 350, 75] as const;
+const WIKI_POSITION = [100, 510, 350, 75] as const;
 const WIKI_LINK = "https://github.com/MayaTheFoxy/MPA/wiki";
-const RESET_POSITION = [
+const DISCORD_INVITE = "https://discord.gg/9GHwmHEMAS";
+const DISCORD_POSITION = [
     WIKI_POSITION[0],
     WIKI_POSITION[1] + WIKI_POSITION[3] + BUTTON_GAP,
     WIKI_POSITION[2],
     WIKI_POSITION[3]
+] as const;
+const RESET_POSITION = [
+    DISCORD_POSITION[0],
+    DISCORD_POSITION[1] + DISCORD_POSITION[3] + BUTTON_GAP,
+    DISCORD_POSITION[2],
+    DISCORD_POSITION[3]
 ] as const;
 const ALERT_POSITION = [
     RESET_POSITION[0],
@@ -62,6 +69,11 @@ const EXPORT_POSITION = [
 ] as const;
 const RESET_CONFIRM_POSITION = [300, 700, 250, 90] as const;
 const RESET_CANCEL_POSITION = [1450, 700, 250, 90] as const;
+const VERSION_TEXT = [
+    WIKI_POSITION[2] / 2 + WIKI_POSITION[0],
+    WIKI_POSITION[1] - (BUTTON_GAP * 2),
+    WIKI_POSITION[2]
+] as const;
 
 // Can add or remove modules depending on if wanting to display them or not
 // Order given is the order displayed
@@ -226,9 +238,9 @@ function DrawSubMenuOptions(subMenu: ModuleTitle): void
           || (
               !Player.CanInteract() && (
                   (settingChar?.MemberNumber === Player.MemberNumber
-                  && !Player.MPA[ModuleTitle.Authority].boundAccessSelf)
+                    && !Player.MPA[ModuleTitle.Authority].boundAccessSelf)
                   || (settingChar?.MemberNumber !== Player.MemberNumber
-                  && !Player.MPA[ModuleTitle.Authority].boundAccessOthers)
+                    && !Player.MPA[ModuleTitle.Authority].boundAccessOthers)
               )
           );
 
@@ -341,9 +353,9 @@ function GetClickedOption(subMenu: ModuleTitle): void
           || (
               !Player.CanInteract() && (
                   (settingChar?.MemberNumber === Player.MemberNumber
-                  && !Player.MPA[ModuleTitle.Authority].boundAccessSelf)
+                    && !Player.MPA[ModuleTitle.Authority].boundAccessSelf)
                   || (settingChar?.MemberNumber !== Player.MemberNumber
-                  && !Player.MPA[ModuleTitle.Authority].boundAccessOthers)
+                    && !Player.MPA[ModuleTitle.Authority].boundAccessOthers)
               )
           );
 
@@ -381,11 +393,11 @@ function GetClickedOption(subMenu: ModuleTitle): void
                 )
                 && !disabledSetting
                 && (setting.loop
-                || index > 0)
+                  || index > 0)
                 // Can't decrease authority beyond if the player is not of that rank
                 && (subMenu !== ModuleTitle.Authority
-                || !AUTHORITY_GROUP_OPTIONS.includes(settingChar!.MPA[subMenu][settingName])
-                || AuthorityIsComparisonToCharacter(settingChar!.MPA[subMenu][settingName] as AuthorityGroup, ">=", Player.MemberNumber ?? -1, settingChar!))
+                  || !AUTHORITY_GROUP_OPTIONS.includes(settingChar!.MPA[subMenu][settingName])
+                  || AuthorityIsComparisonToCharacter(settingChar!.MPA[subMenu][settingName] as AuthorityGroup, ">=", Player.MemberNumber ?? -1, settingChar!))
             )
             {
                 const prevValue = settingChar!.MPA[subMenu][settingName];
@@ -406,11 +418,11 @@ function GetClickedOption(subMenu: ModuleTitle): void
                 )
                 && !disabledSetting
                 && (setting.loop
-                || index + 1 < setting.options.length)
+                  || index + 1 < setting.options.length)
                 // Can't increase authority beyond what the player who is setting it is
                 && (subMenu !== ModuleTitle.Authority
-                || !AUTHORITY_GROUP_OPTIONS.includes(settingChar!.MPA[subMenu][settingName])
-                || AuthorityIsComparisonToCharacter(settingChar!.MPA[subMenu][settingName] as AuthorityGroup, ">", Player.MemberNumber ?? -1, settingChar!))
+                  || !AUTHORITY_GROUP_OPTIONS.includes(settingChar!.MPA[subMenu][settingName])
+                  || AuthorityIsComparisonToCharacter(settingChar!.MPA[subMenu][settingName] as AuthorityGroup, ">", Player.MemberNumber ?? -1, settingChar!))
             )
             {
                 const prevValue = settingChar!.MPA[subMenu][settingName];
@@ -462,6 +474,11 @@ export function PreferenceMenuClick(): void
         if (MouseIn(...WIKI_POSITION))
         {
             window.open(WIKI_LINK, "_blank")?.focus();
+            return;
+        }
+        if (MouseIn(...DISCORD_POSITION))
+        {
+            window.open(DISCORD_INVITE, "_blank")?.focus();
             return;
         }
         if (MouseIn(...RESET_POSITION))
@@ -651,6 +668,13 @@ export function PreferenceMenuRun(): void
             LocalizedText("Open the MPA wiki in a new tab")
         );
         DrawButton(
+            ...DISCORD_POSITION,
+            "Discord",
+            "#ffffff",
+            ICONS.DISCORD,
+            "Join the MPA discord!"
+        );
+        DrawButton(
             ...RESET_POSITION,
             LocalizedText("RESET"),
             "#ff2e2eaa",
@@ -670,6 +694,12 @@ export function PreferenceMenuRun(): void
             "#ffffff",
             "",
             LocalizedText("Export your current settings to the clipboard")
+        );
+        DrawTextFit(
+            `${LocalizedText("Version")}: ${Player.MPA.version}`,
+            ...VERSION_TEXT,
+            "Black",
+            "Gray"
         );
     }